requirements.txt
alembic==1.7.7
ddt==1.4.4
flake8==4.0.1
greenlet==1.1.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mysqlclient==2.1.0
nose==1.3.7
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
SQLAlchemy==1.4.36
zipp==3.8.0
